Lagt til SimuleringMapper og trigger simulering i BehandlingresultatSteg
