+++ src/main/kotlin/no/nav/familie/ks/sak/api/SimuleringController.kt
@@ -2,7 +2,9 @@ package no.nav.familie.ks.sak.api
 
 import no.nav.familie.kontrakter.felles.Ressurs
 import no.nav.familie.ks.sak.api.dto.SimuleringDto
+import no.nav.familie.ks.sak.api.mapper.SimuleringMapper.tilSimuleringDto
 import no.nav.familie.ks.sak.config.BehandlerRolle
+import no.nav.familie.ks.sak.kjerne.behandling.steg.simulering.SimuleringService
 import no.nav.familie.ks.sak.sikkerhet.AuditLoggerEvent
 import no.nav.familie.ks.sak.sikkerhet.TilgangService
 import no.nav.security.token.support.core.api.ProtectedWithClaims
@@ -30,8 +32,10 @@ class SimuleringController(
             minimumBehandlerRolle = BehandlerRolle.VEILEDER,
             handling = "hente og/eller oppdatere simulering på behandling"
         )
-        val vedtakSimuleringMottaker = simuleringService.oppdaterSimuleringPåBehandlingVedBehov(behandlingId)
-        val restSimulering = vedtakSimuleringMottakereTilRestSimulering(vedtakSimuleringMottaker)
-        return ResponseEntity.ok(Ressurs.success(restSimulering))
+        return ResponseEntity.ok(
+            Ressurs.success(
+                simuleringService.oppdaterSimuleringPåBehandlingVedBehov(behandlingId).tilSimuleringDto()
+            )
+        )
     }
 }
